feat(genetic): add v2 genetic loop

+++ apps/panacee-genetics/src/datas/fitness.h
@@ -3,9 +3,9 @@ typedef struct
     int hospital_count;
     int uhc_count;
     int distant_resident_count;
-    int distant_resident_percent;
+    double distant_resident_percent;
     int distant_town_count;
-    int distant_town_percent;
+    double distant_town_percent;
     double fitness_score;
     int fitness_count;
     double fitness_average;
